requirements.txt

# === Core Google Sheets stack ===
gspread>=6.0.0
google-auth>=2.28.0
requests>=2.31.0

# === Desktop UI ===
PyQt5>=5.15.11

# === Config & secrets ===
python-dotenv>=1.0.1
pyzipper>=0.3.6

# ============================================================================
# IMPROVEMENTS - Added by automatic assembler
# ============================================================================

# Security
cryptography>=41.0.0
keyring>=24.0.0

# CLI utility
click>=8.1.0

# Database encryption (choose one)
# sqlcipher3-binary>=0.5.0  # Опционально, для шифрования БД
# sqlcipher3>=0.5.0        # For Linux/Mac
